refactor(score): extract the shared Prometheus exposition parser
